+++ crypto/fipsmodule/bcm.c
@@ -142,7 +142,6 @@
 #include "pqdsa/pqdsa.c"
 #include "rand/ctrdrbg.c"
 #include "rand/rand.c"
-#include "rand/urandom.c"
 #include "rand/entropy/entropy_sources.c"
 #include "rand/entropy/tree_drbg_jitter_entropy.c"
 #include "rsa/blinding.c"
