+++ crypto/rand_extra/getentropy.c
@@ -8,7 +8,7 @@
 
 #include <openssl/rand.h>
 
-#include "../fipsmodule/rand/internal.h"
+#include "internal.h"
 
 #if defined(OPENSSL_RAND_GETENTROPY)
