+++ crypto/rand_extra/getentropy_test.cc
@@ -18,7 +18,7 @@
 
 #include <openssl/rand.h>
 
-#include "../fipsmodule/rand/internal.h"
+#include "internal.h"
 
 #if defined(OPENSSL_RAND_GETENTROPY)
