+++ crypto/rand_extra/urandom.c
@@ -73,8 +73,7 @@
 #include <openssl/mem.h>
 
 #include "getrandom_fillin.h"
-#include "../delocate.h"
-#include "../../internal.h"
+#include "../internal.h"
 
 #ifndef MIN
 #define AWSLC_MIN(X,Y) (((X) < (Y)) ? (X) : (Y))
@@ -169,17 +168,17 @@ static ssize_t boringssl_getrandom(void *buf, size_t buf_len, unsigned flags) {
 static const int kHaveGetrandom = -3;
 
 // urandom_fd is a file descriptor to /dev/urandom. It's protected by |once|.
-DEFINE_BSS_GET(int, urandom_fd)
+static int urandom_fd = 0;
 
 #if defined(USE_NR_getrandom)
 
 // getrandom_ready is one if |getrandom| had been initialized by the time
 // |init_once| was called and zero otherwise.
-DEFINE_BSS_GET(int, getrandom_ready)
+static int getrandom_ready = 0;
 
 // extra_getrandom_flags_for_seed contains a value that is ORed into the flags
 // for getrandom() when reading entropy for a seed.
-DEFINE_BSS_GET(int, extra_getrandom_flags_for_seed)
+static int extra_getrandom_flags_for_seed = 0;
 
 // On Android, check a system property to decide whether to set
 // |extra_getrandom_flags_for_seed| otherwise they will default to zero.  If
@@ -196,14 +195,14 @@ static void maybe_set_extra_getrandom_flags(void) {
 
   value[length] = 0;
   if (OPENSSL_strcasecmp(value, "true") == 0) {
-    *extra_getrandom_flags_for_seed_bss_get() = GRND_RANDOM;
+    extra_getrandom_flags_for_seed = GRND_RANDOM;
   }
 #endif
 }
 
 #endif  // USE_NR_getrandom
 
-DEFINE_STATIC_ONCE(rand_once)
+static CRYPTO_once_t urandom_initialize_once = CRYPTO_ONCE_INIT;
 
 // init_once initializes the state of this module to values previously
 // requested. This is the only function that modifies |urandom_fd|, which may be
@@ -215,7 +214,7 @@ static void init_once(void) {
   ssize_t getrandom_ret =
       boringssl_getrandom(&dummy, sizeof(dummy), GRND_NONBLOCK);
   if (getrandom_ret == 1) {
-    *getrandom_ready_bss_get() = 1;
+    getrandom_ready = 1;
     have_getrandom = 1;
   } else if (getrandom_ret == -1 && errno == EAGAIN) {
     // We have getrandom, but the entropy pool has not been initialized yet.
@@ -230,7 +229,7 @@ static void init_once(void) {
   }
 
   if (have_getrandom) {
-    *urandom_fd_bss_get() = kHaveGetrandom;
+    urandom_fd = kHaveGetrandom;
     maybe_set_extra_getrandom_flags();
     return;
   }
@@ -272,19 +271,19 @@ static void init_once(void) {
       abort();
     }
   }
-  *urandom_fd_bss_get() = fd;
+  urandom_fd = fd;
 }
 
-DEFINE_STATIC_ONCE(wait_for_entropy_once)
+static CRYPTO_once_t wait_for_entropy_once = CRYPTO_ONCE_INIT;
 
 static void wait_for_entropy(void) {
-  int fd = *urandom_fd_bss_get();
+  int fd = urandom_fd;
   if (fd == kHaveGetrandom) {
     // |getrandom| and |getentropy| support blocking in |fill_with_entropy|
     // directly. For |getrandom|, we first probe with a non-blocking call to aid
     // debugging.
 #if defined(USE_NR_getrandom)
-    if (*getrandom_ready_bss_get()) {
+    if (getrandom_ready) {
       // The entropy pool was already initialized in |init_once|.
       return;
     }
@@ -355,7 +354,7 @@ static void wait_for_entropy(void) {
 // until the entropy pool is initialized. Otherwise, this function may fail,
 // setting |errno| to |EAGAIN| if the entropy pool has not yet been initialized.
 // If |seed| is one, this function will OR in the value of
-// |*extra_getrandom_flags_for_seed()| when using |getrandom|.
+// |extra_getrandom_flags_for_seed| when using |getrandom|.
 static int fill_with_entropy(uint8_t *out, size_t len, int block, int seed) {
   if (len == 0) {
     return 1;
@@ -381,13 +380,13 @@ static int fill_with_entropy(uint8_t *out, size_t len, int block, int seed) {
 
 #if defined (USE_NR_getrandom)
   if (seed) {
-    getrandom_flags |= *extra_getrandom_flags_for_seed_bss_get();
+    getrandom_flags |= extra_getrandom_flags_for_seed;
   }
 #endif
 
   CRYPTO_init_sysrand();
   if (block) {
-    CRYPTO_once(wait_for_entropy_once_bss_get(), wait_for_entropy);
+    CRYPTO_once(&wait_for_entropy_once, wait_for_entropy);
   }
 
   // Clear |errno| so it has defined value if |read| or |getrandom|
@@ -396,7 +395,7 @@ static int fill_with_entropy(uint8_t *out, size_t len, int block, int seed) {
   while (len > 0) {
     ssize_t r;
 
-    if (*urandom_fd_bss_get() == kHaveGetrandom) {
+    if (urandom_fd == kHaveGetrandom) {
 #if defined(USE_NR_getrandom)
       r = boringssl_getrandom(out, len, getrandom_flags);
 #else  // USE_NR_getrandom
@@ -407,7 +406,7 @@ static int fill_with_entropy(uint8_t *out, size_t len, int block, int seed) {
       size_t retry_counter = 0;
       long backoff = INITIAL_BACKOFF_DELAY;
       do {
-        r = read(*urandom_fd_bss_get(), out, len);
+        r = read(urandom_fd, out, len);
         if ((r == -1) && (errno != EINTR)) {
           if (retry_counter >= MAX_BACKOFF_RETRIES) {
             break;
@@ -433,7 +432,7 @@ static int fill_with_entropy(uint8_t *out, size_t len, int block, int seed) {
 }
 
 void CRYPTO_init_sysrand(void) {
-  CRYPTO_once(rand_once_bss_get(), init_once);
+  CRYPTO_once(&urandom_initialize_once, init_once);
 }
 
 // CRYPTO_sysrand puts |requested| random bytes into |out|.
