+++ crypto/rand_extra/windows.c
@@ -14,7 +14,7 @@
 
 #include <openssl/rand.h>
 
-#include "../fipsmodule/rand/internal.h"
+#include "internal.h"
 
 #if defined(OPENSSL_RAND_WINDOWS)
